+++ src/modules/Projects/Projects.container.tsx
@@ -1,13 +1,13 @@
 import * as React from 'react'
 import styled from 'styled-components'
-import { ProjectCard } from './ProjectCard'
-import { ProjectsHero } from './ProjectsHero'
+import { ProjectCard } from './components/ProjectCard'
+import { ProjectsHero } from './components/ProjectsHero'
 import { Spinner } from '../../components/common/Spinner'
 import { connect } from 'react-redux'
 import { RootState } from '../../common/redux/types'
 import { contentType } from '../../types/models'
 import { UserInfo } from '../account/types'
-import { ProjectsDashboard } from './ProjectsDashboard'
+import { ProjectsDashboard } from './components/ProjectsDashboard'
 import FilterSortButtons from '../../components/common/AdvancedFilter/FilterSortButtons'
 import { deviceWidth } from '../../lib/commonData'
 import { fetchProjects } from './Projects.actions'
