+++ src/modules/Projects/components/ProjectsHero.tsx
@@ -1,12 +1,12 @@
 import * as React from 'react'
 import styled from 'styled-components'
-import HeaderSubTabs from '../../components/common/HeaderSubTabs'
-import { SingleStatistic } from '../../components/common/SingleStatistic'
-import { StatType } from '../../types/models'
-import { deviceWidth } from '../../lib/commonData'
-import Search from '../../components/common/Search'
-import { contentType } from '../../types/models'
-import * as instanceSettings from '../../instance-settings'
+import HeaderSubTabs from '../../../components/common/HeaderSubTabs'
+import { SingleStatistic } from '../../../components/common/SingleStatistic'
+import { StatType } from '../../../types/models'
+import { deviceWidth } from '../../../lib/commonData'
+import Search from '../../../components/common/Search'
+import { contentType } from '../../../types/models'
+import * as instanceSettings from '../../../instance-settings'
 
 const ContainerInner = styled.div`
   height: auto;
